### 2021-02-20 #3046 @whedon
[PRE REVIEW]: imodels: a python package for fitting interpretable models
**Submitting author:** @csinva (<a href="http://orcid.org/0000-0003-0318-2340">Chandan Singh</a>)
**Repository:** <a href="https://github.com/csinva/imodels" target ="_blank">https://github.com/csinva/imodels</a>
**Version:** v0.2.7
