+++ docs/paper/references.bib
@@ -1,96 +1,108 @@
 @article{ustun2016supersparse,
-  title={Supersparse linear integer models for optimized medical scoring systems},
-  author={Ustun, Berk and Rudin, Cynthia},
-  journal={Machine Learning},
-  volume={102},
-  number={3},
-  pages={349--391},
-  year={2016},
-  publisher={Springer}
+    title = {Supersparse linear integer models for optimized medical scoring systems},
+    author = {Ustun, Berk and Rudin, Cynthia},
+    journal = {Machine Learning},
+    volume = {102},
+    number = {3},
+    pages = {349--391},
+    year = {2016},
+    publisher = {Springer},
+    doi = {10.1007/s10994-015-5528-6},
 }
 
 @article{pedregosa2011scikit,
-  title={Scikit-learn: Machine learning in Python},
-  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
-  journal={the Journal of machine Learning research},
-  volume={12},
-  pages={2825--2830},
-  year={2011},
-  publisher={JMLR. org}
+    title = {Scikit-learn: Machine learning in Python},
+    author = {Pedregosa, Fabian and Varoquaux, Ga {\"e} l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
+    journal = {the Journal of machine Learning research},
+    volume = {12},
+    pages = {2825--2830},
+    year = {2011},
+    publisher = {JMLR. org},
+    url = {http://jmlr.org/papers/v12/pedregosa11a.html}
 }
 
 @article{letham2015interpretable,
-  title={Interpretable classifiers using rules and bayesian analysis: Building a better stroke prediction model},
-  author={Letham, Benjamin and Rudin, Cynthia and McCormick, Tyler H and Madigan, David and others},
-  journal={Annals of Applied Statistics},
-  volume={9},
-  number={3},
-  pages={1350--1371},
-  year={2015},
-  publisher={Institute of Mathematical Statistics}
+    title = {Interpretable classifiers using rules and bayesian analysis: Building a better stroke prediction model},
+    author = {Letham, Benjamin and Rudin, Cynthia and McCormick, Tyler H and Madigan, David and others},
+    journal = {Annals of Applied Statistics},
+    volume = {9},
+    number = {3},
+    pages = {1350--1371},
+    year = {2015},
+    publisher = {Institute of Mathematical Statistics},
+    doi = {10.1214/15-aoas848},
 }
 
 @article{holte1993very,
-  title={Very simple classification rules perform well on most commonly used datasets},
-  author={Holte, Robert C},
-  journal={Machine learning},
-  volume={11},
-  number={1},
-  pages={63--90},
-  year={1993},
-  publisher={Springer}
+    title = {Very simple classification rules perform well on most commonly used datasets},
+    author = {Holte, Robert C},
+    journal = {Machine learning},
+    volume = {11},
+    number = {1},
+    pages = {63--90},
+    year = {1993},
+    publisher = {Springer},
+    doi = {10.1023/A:1022631118932}
 }
 
 @book{breiman1984classification,
-  title={Classification and regression trees},
-  author={Breiman, Leo and Friedman, Jerome and Stone, Charles J and Olshen, Richard A},
-  year={1984},
-  publisher={CRC press}
[…]
