### 2021-02-22 f448d78
added complexity to comparison notebook
