+++ imodels/rule_set/skope_rules.py
@@ -218,7 +218,7 @@ class SkopeRulesClassifier(BaseEstimator, RuleSet):
 
     def __init__(self,
                  precision_min=0.5,
-                 recall_min=0.01,
+                 recall_min=0.4,
                  n_estimators=10,
                  max_samples=.8,
                  max_samples_features=1.,
