### 2021-02-22 c09bbae
add comparison notebook as test
