### 2021-02-22 2f3e3cd
add tqdm to reqs
