### 2021-03-07 fc7edb5
added slimclassifier
