### 2021-03-07 20c59e1
added slimclassifier to demo
