### 2021-03-07 b8c8155
add slimclassifier to docs
