### 2021-03-09 98e9a09
remove unpenalized lasso from demo
