+++ tests/notebooks/imodels_demo.py
@@ -243,7 +243,7 @@ y_sim = 1 * X_sim[:, 0] + 2 * X_sim[:, 1] - 1 * X_sim[:, 2] + np.random.randn(n)
 # fit linear models with different regularization parameters
 print('groundtruth weights should be 1, 2, -1...')
 model = SLIMRegressor()
-for lambda_reg in [0, 1e-2, 5e-2, 1e-1, 1, 2, 5, 10]:
+for lambda_reg in [1e-3, 1e-2, 5e-2, 1e-1, 1, 2, 5, 10]:
     model.fit(X_sim, y_sim, lambda_reg)
     mse = np.mean(np.square(y_sim - model.predict(X_sim)))
     print(f'lambda: {lambda_reg}\tmse: {mse: 0.2f}\tweights: {model.model.coef_}')
@@ -255,7 +255,7 @@ y_sim = np.round(y_sim)
 # fit linear models with different regularization parameters
 print('groundtruth weights should be 1, 2, -1...')
 model = SLIMClassifier()
-for lambda_reg in [1e-2, 5e-2, 1e-1, 1, 2, 5, 10]:
+for lambda_reg in [1e-3, 1e-2, 5e-2, 1e-1, 1, 2, 5, 10]:
     model.fit(X_sim, y_sim, lambda_reg)
     mll = np.mean(metrics.log_loss(y_sim, model.predict(X_sim)))
     print(f'lambda: {lambda_reg}\tmlogloss: {mll: 0.2f}\tweights: {model.model.coef_}')
