### 2021-03-09 e84307a
brl one-hot strings by default
