### 2021-03-10 129c9a1
update slim docstring
