### 2021-03-11 4191953
fit should return self
