+++ tests/estimator_checks_test.py
@@ -1,0 +1,17 @@
+import numpy as np
+import unittest
+from sklearn.utils.estimator_checks import check_estimator
+from imodels import *
+
+"""
+class TestCheckEstimators(unittest.TestCase):
+    '''Checks that estimators conform to sklearn checks
+    '''
+
+    def test_classification_binary(self):
+        '''Test imodels on basic binary classification task
+        '''
+        for classifier in [SLIMClassifier]:
+            print(classifier)
+            assert check_estimator(classifier())
+"""
