+++ tests/test_notebooks.py
@@ -8,6 +8,6 @@ import matplotlib.pyplot
 def test_imodels_demo(mock_pd_style, mock_plot_tree, mock_pyplot):
     from .notebooks import imodels_demo
 
-@patch('matplotlib.pyplot')
-def test_imodels_comparisons(mock_pyplot):
-    from .notebooks import imodels_comparisons
+# @patch('matplotlib.pyplot')
+# def test_imodels_comparisons(mock_pyplot):
+#     from .notebooks import imodels_comparisons

[…]
