### 2021-03-11 bbc1970
boosted rules compat
