### 2021-03-11 d56ccbb
add argument checks
