+++ imodels/rule_list/greedy_rule_list.py
@@ -4,12 +4,14 @@
 import math
 import numpy as np
 from copy import deepcopy
-from sklearn.base import BaseEstimator
+from sklearn.base import BaseEstimator, ClassifierMixin, RegressorMixin
+from sklearn.utils.multiclass import check_classification_targets
+from sklearn.utils.validation import check_X_y, check_array, check_is_fitted
 
 from imodels.rule_list.rule_list import RuleList
 
 
-class GreedyRuleListClassifier(BaseEstimator, RuleList):
+class GreedyRuleListClassifier(BaseEstimator, RuleList, ClassifierMixin):
     def __init__(self, max_depth: int = 5, class_weight=None, criterion: str = 'gini', strategy: str = 'max'):
         '''
         Params
