+++ imodels/rule_list/greedy_rule_list.py
@@ -27,7 +27,6 @@ class GreedyRuleListClassifier(BaseEstimator, RuleList, ClassifierMixin):
         '''
 
         self.max_depth = max_depth
-        self.feature_names = None
         self.class_weight = class_weight
         self.criterion = criterion
         self.strategy = strategy
@@ -50,13 +49,9 @@ class GreedyRuleListClassifier(BaseEstimator, RuleList, ClassifierMixin):
             x = x.values
         else:
             if feature_names is None:
-                self.feature_names = ['feat ' + str(i) for i in range(x.shape[1])]
+                self.feature_names_ = ['feat ' + str(i) for i in range(x.shape[1])]
         if feature_names is not None:
-            self.feature_names = feature_names
-        if 'pandas' in str(type(y)):
-            y = y.values
-        assert type(x) == np.ndarray, 'x is not numpy array'
-        assert type(y) == np.ndarray, 'y is not numpy array'
+            self.feature_names_ = feature_names
 
         # base case 1: no data in this group
         if len(y) == 0:
@@ -95,11 +90,11 @@ class GreedyRuleListClassifier(BaseEstimator, RuleList, ClassifierMixin):
             # print
             if verbose:
                 print(
-                    f'{np.mean(100 * y):.2f} -> {self.feature_names[col]} -> {np.mean(100 * y_left):.2f} ({y_left.size}) {np.mean(100 * y_right):.2f} ({y_right.size})')
+                    f'{np.mean(100 * y):.2f} -> {self.feature_names_[col]} -> {np.mean(100 * y_left):.2f} ({y_left.size}) {np.mean(100 * y_right):.2f} ({y_right.size})')
 
             # save info
             par_node = [{
-                'col': self.feature_names[col],
+                'col': self.feature_names_[col],
                 'index_col': col,
                 'cutoff': cutoff,
                 'val': np.mean(y),  # values before splitting
@@ -114,12 +109,12 @@ class GreedyRuleListClassifier(BaseEstimator, RuleList, ClassifierMixin):
 
             self.depth += 1  # increase the depth since we call fit once
             self.rules_ = par_node
-            self.complexity = len(self.rules_)
+            self.complexity_ = len(self.rules_)
             return par_node
 
     def predict_proba(self, X):
-        if 'pandas' in str(type(X)):
-            X = X.values
+        check_is_fitted(self)
+        X = check_array(X)
         n = X.shape[0]
         probs = np.zeros(n)
         for i in range(n):
@@ -133,6 +128,8 @@ class GreedyRuleListClassifier(BaseEstimator, RuleList, ClassifierMixin):
         return np.vstack((1 - probs, probs)).transpose()  # probs (n, 2)
 
     def predict(self, X):
+        check_is_fitted(self)
+        X = check_array(X)
         return np.argmax(self.predict_proba(X), axis=1)
 
     def __str__(self):
