+++ imodels/rule_list/one_r.py
@@ -51,7 +51,7 @@ class OneRClassifier(GreedyRuleListClassifier):
             # print('acc', feat_names_single[0], f'{accs[col_idx]:0.2f}')
         col_idx_best = np.argmax(accs)
         self.rules_ = ms[col_idx_best].rules_
-        self.complexity = len(self.rules_)
+        self.complexity_ = len(self.rules_)
 
         # need to adjust index_col since was fitted with only 1 col
         for rule in self.rules_:
