### 2021-03-11 fd5734e
dont install tests
