### 2021-03-11 c00774d
ROCAUC + val set + move comparison running to script
