+++ tests/notebooks/imodels_comparisons.py
@@ -6,7 +6,7 @@
 #       extension: .py
 #       format_name: percent
 #       format_version: '1.3'
-#       jupytext_version: 1.10.2
+#       jupytext_version: 1.10.3
 #   kernelspec:
 #     display_name: Python 3
 #     language: python
@@ -14,16 +14,17 @@
 # ---
 
 # %%
-import pickle as pkl
-import pandas as pd
-import imodels
 import itertools
 import os
-from imodels.util.evaluate.compare_models import run_comparison
-from sklearn.metrics import accuracy_score, f1_score
-from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
-import matplotlib.pyplot as plt
+import pickle as pkl
+
+import numpy as np
+import pandas as pd
 from tqdm import tqdm
+import matplotlib.pyplot as plt
+from sklearn.metrics import accuracy_score, roc_auc_score
+from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
+from sklearn.datasets import fetch_openml
 
 # %load_ext autoreload
 # %autoreload 2
@@ -33,20 +34,9 @@ if os.getcwd().split('/')[-1] != 'imodels':
     os.chdir('..')
 
 MODEL_COMPARISON_PATH = 'tests/test_data/comparison_data/'
-MODEL_COMPARISON_FILE = MODEL_COMPARISON_PATH + 'model_comparisons.pkl'
-
-# %% [markdown]
-# # compare static performance of different models
 
 # %%
-df = pkl.load(open(MODEL_COMPARISON_FILE, 'rb'))['df'].round(3)
-print('columns', df.columns, 'models', df.index)
-
-# %% [markdown]
-# # complexity-vs-accuracy plots for each model
-
-# %%
-COMPARISON_DATASETS = [
+datasets = [
         ("breast-cancer", 13),
         ("breast-w", 15),
         ("credit-g", 31),
@@ -56,85 +46,114 @@ COMPARISON_DATASETS = [
         ("vote", 56),
     ]
 
-METRICS = [
-    ('Acc.', accuracy_score),
-    ('Time', None),
-    ('Complexity', None)
-    
-]
-
-
-def get_comparison_df(estimators):
-    '''Get results for running multiple estimators
-    '''
-    estimator_name = estimators[0][0]
-    model_comparison_file = MODEL_COMPARISON_PATH + f'{estimator_name}_comparisons.pkl'
-    if os.path.isfile(model_comparison_file):
-        result = pkl.load(open(model_comparison_file, 'rb'))['df']
+def get_comparison_result(path, estimator_name, test=False):
+    if test:
+        result_file = path + f'{estimator_name}_test_comparisons.pkl'
     else:
-        result = run_comparison(COMPARISON_DATASETS, METRICS, estimators, write=False, average=True, verbose=False)
-        pkl.dump({'df': result}, open(model_comparison_file, 'wb'))
-    return result
+        result_file = path + f'{estimator_name}_comparisons.pkl'
+    return pkl.load(open(result_file, 'rb'))
 
-def viz_model(result):
-    '''Plot acc vs complexity
+def collect_result_files(path, test=False):
+    if test:
+        fn = lambda x: '_test_' in x
+    else:
+        fn = lambda x: '_test_' not in x
+    return list(filter(fn, os.listdir(path)))
+
+def viz_comparison(result_df, result_estimators, dpi=83):
+    '''Plot ROC AUC vs complexity
     '''
-    complexities = result[result.index.str.contains('Complexity')]
[…]
