+++ tests/compare_models.py
@@ -1,0 +1,296 @@
+'''Compare different estimators on public datasets
+Code modified from https://github.com/tmadl/sklearn-random-bits-forest
+'''
+import os 
+import time
+import re, string
+import pickle as pkl
+import argparse
+import logging
+
+import numpy as np
+from sklearn.ensemble import RandomForestClassifier
+from sklearn.ensemble import GradientBoostingClassifier
+from sklearn.neural_network import MLPClassifier
+from sklearn.metrics import accuracy_score, roc_auc_score
+from sklearn.model_selection import KFold, train_test_split
+from sklearn.datasets import fetch_openml
+from sklearn.preprocessing import OneHotEncoder
+from scipy.stats.stats import mannwhitneyu, ttest_ind
+from tqdm import tqdm
+import pandas as pd
+import imodels
+
+
+COMPARISON_DATASETS = [
+    ("breast-cancer", 13),
+    ("breast-w", 15),
+    ("credit-g", 31),
+    ("haberman", 43),
+    ("heart", 1574),
+    ("labor", 4),
+    ("vote", 56),
+]
+
+METRICS = [
+    ('Accuracy', accuracy_score),
+    ('ROC Score', roc_auc_score),
+    ('Time', None),
+    ('Complexity', None)
+]
+
+BEST_ESTIMATORS = [
+    [('random_forest', RandomForestClassifier(n_estimators=n, max_depth=2)) for n in np.linspace(2, 8, 5, dtype=int)],
+    [('gradient_boosting', GradientBoostingClassifier(n_estimators=n, max_depth=2)) for n in np.linspace(2, 8, 5, dtype=int)],
+    [('skope_rules', imodels.SkopeRulesClassifier(n_estimators=n, max_depth=2)) for n in np.linspace(2, 200, 5, dtype=int)],
+    [('rulefit', imodels.RuleFitClassifier(max_rules=n, tree_size=2)) for n in np.linspace(2, 100, 5, dtype=int)],
+    [('fplasso', imodels.FPLassoClassifier(max_rules=n, maxcardinality=2)) for n in np.linspace(2, 60, 5, dtype=int)],
+    [('fpskope', imodels.FPSkopeClassifier(maxcardinality=n, max_depth_duplication=3)) for n in np.arange(1, 5)],
+    [('brl', imodels.BayesianRuleListClassifier(listlengthprior=n, maxcardinality=2)) for n in [2, 4, 8, 16]],
+    [('grl', imodels.GreedyRuleListClassifier(max_depth=n)) for n in np.arange(1, 6)],
+    [('oner', imodels.OneRClassifier(max_depth=n)) for n in np.arange(1, 6)],
+    [('brs', imodels.BoostedRulesClassifier(n_estimators=n)) for n in np.linspace(2, 32, 5, dtype=int)]
+]
+
+ALL_ESTIMATORS = []
+ALL_ESTIMATORS.append(
+    [('random_forest - depth_1', RandomForestClassifier(n_estimators=n, max_depth=1)) for n in np.linspace(2, 40, 5, dtype=int)]
+    + [('random_forest - depth_2', RandomForestClassifier(n_estimators=n, max_depth=2)) for n in np.linspace(2, 15, 5, dtype=int)]
+    + [('random_forest - depth_3', RandomForestClassifier(n_estimators=n, max_depth=3)) for n in np.arange(2, 8)]
+)
+ALL_ESTIMATORS.append(
+    [('gradient_boosting - depth_1', GradientBoostingClassifier(n_estimators=n, max_depth=1)) for n in np.linspace(2, 40, 5, dtype=int)]
[…]
