### 2021-03-11 19a530f
update comparison cache
