### 2021-03-11 571fff5
update setup.py descr
