+++ readme.md
@@ -14,7 +14,7 @@
   <a href="https://github.com/csinva/imodels/actions"><img src="https://github.com/csinva/imodels/workflows/tests/badge.svg"></a>
   <img src="https://img.shields.io/github/checks-status/csinva/imodels/master">
   <img src="https://img.shields.io/pypi/v/imodels?color=orange">
-  <img src="https://static.pepy.tech/personalized-badge/imodels?period=total&units=none&left_color=gray&right_color=orange&left_text=downloads&kill_cache=0">
+  <img src="https://static.pepy.tech/personalized-badge/imodels?period=total&units=none&left_color=gray&right_color=orange&left_text=downloads&kill_cache=2">
 </p>  
 
 
@@ -100,7 +100,6 @@ Demos are contained in the [notebooks](notebooks) folder.
 > [imodels_demo](notebooks/imodels_demo.ipynb) - shows how to fit, predict, and visualize with different interpretable models
 >
 > [clinical decision rule notebook](https://github.com/csinva/iai-clinical-decision-rule/blob/master/notebooks/05_fit_interpretable_models.ipynb) - shows an example of using `imodels` for deriving a clinical decision rule
->
 > we also include some demos of posthoc analysis, which occurs after fitting models
 >
 >  - [posthoc.ipynb](notebooks/2_posthoc.ipynb) - shows different simple analyses to interpret a trained model
