### 2021-03-14 dc2326f
simplify viz_comparison
