### 2021-03-15 fe3ad55
compute auc-of-auc in compare_models
