+++ tests/compare_models.py
@@ -1,20 +1,23 @@
 '''Compare different estimators on public datasets
 Code modified from https://github.com/tmadl/sklearn-random-bits-forest
 '''
-import os
-import time
-import pickle as pkl
 import argparse
+import os
+import pickle as pkl
+import time
+from typing import Dict, Any
 
+import imodels
 import numpy as np
-from sklearn.ensemble import RandomForestClassifier
+import pandas as pd
+from scipy.interpolate import interp1d
+from scipy.sparse import issparse
+from sklearn.datasets import fetch_openml
 from sklearn.ensemble import GradientBoostingClassifier
+from sklearn.ensemble import RandomForestClassifier
 from sklearn.metrics import accuracy_score, roc_auc_score
 from sklearn.model_selection import KFold, train_test_split
-from sklearn.datasets import fetch_openml
 from tqdm import tqdm
-import pandas as pd
-import imodels
 
 
 COMPARISON_DATASETS = [
@@ -28,11 +31,17 @@ COMPARISON_DATASETS = [
 ]
 
 METRICS = [
-    ('Accuracy', accuracy_score),
-    ('ROC Score', roc_auc_score),
-    ('Time', None),
-    ('Complexity', None)
+    ('accuracy', accuracy_score),
+    ('ROCAUC', roc_auc_score),
+    ('time', None),
+    ('complexity', None)
 ]
+
+# complexity score under which a model is considered interpretable
+LOW_COMPLEXITY_CUTOFF = 30
+
+# min complexity of curves included in the AUC-of-AUC comparison must be below this value
+MAX_START_COMPLEXITY = 10
 
 BEST_ESTIMATORS = [
     [('random_forest', RandomForestClassifier(n_estimators=n, max_depth=2)) for n in np.linspace(2, 8, 5, dtype=int)],
@@ -49,29 +58,29 @@ BEST_ESTIMATORS = [
 
 ALL_ESTIMATORS = []
 ALL_ESTIMATORS.append(
-    [('random_forest - depth_1', RandomForestClassifier(n_estimators=n, max_depth=1)) for n in np.linspace(2, 40, 5, dtype=int)]
-    + [('random_forest - depth_2', RandomForestClassifier(n_estimators=n, max_depth=2)) for n in np.linspace(2, 15, 5, dtype=int)]
-    + [('random_forest - depth_3', RandomForestClassifier(n_estimators=n, max_depth=3)) for n in np.arange(2, 8)]
+    [('random_forest - depth_1', RandomForestClassifier(n_estimators=n, max_depth=1)) for n in np.linspace(1, 40, 10, dtype=int)]
+    + [('random_forest - depth_2', RandomForestClassifier(n_estimators=n, max_depth=2)) for n in np.linspace(1, 15, 10, dtype=int)]
+    + [('random_forest - depth_3', RandomForestClassifier(n_estimators=n, max_depth=3)) for n in np.arange(1, 8)]
 )
 ALL_ESTIMATORS.append(
-    [('gradient_boosting - depth_1', GradientBoostingClassifier(n_estimators=n, max_depth=1)) for n in np.linspace(2, 40, 5, dtype=int)]
-    + [('gradient_boosting - depth_2', GradientBoostingClassifier(n_estimators=n, max_depth=2)) for n in np.linspace(2, 15, 5, dtype=int)]
-    + [('gradient_boosting - depth_3', GradientBoostingClassifier(n_estimators=n, max_depth=3)) for n in np.arange(2, 8)]
+    [('gradient_boosting - depth_1', GradientBoostingClassifier(n_estimators=n, max_depth=1)) for n in np.linspace(1, 40, 10, dtype=int)]
[…]
