### 2021-03-15 d631030
cache and plot auc-of-auc
