+++ tests/notebooks/imodels_comparisons.py
@@ -17,6 +17,7 @@
 import itertools
 import os
 import pickle as pkl
+from typing import List, Dict, Any
 
 import numpy as np
 import pandas as pd
@@ -33,7 +34,7 @@ from sklearn.datasets import fetch_openml
 if os.getcwd().split('/')[-1] != 'imodels':
     os.chdir('..')
 
-MODEL_COMPARISON_PATH = 'tests/test_data/comparison_data/'
+MODEL_COMPARISON_PATH = 'tests/comparison_data/'
 
 # %%
 datasets = [
@@ -48,39 +49,50 @@ datasets = [
 
 def get_comparison_result(path, estimator_name, test=False):
     if test:
-        result_file = path + f'{estimator_name}_test_comparisons.pkl'
+        result_file = path + 'test/' + f'{estimator_name}_test_comparisons.pkl'
     else:
-        result_file = path + f'{estimator_name}_comparisons.pkl'
+        result_file = path + 'val/' + f'{estimator_name}_comparisons.pkl'
     return pkl.load(open(result_file, 'rb'))
 
-def get_all_results(path, test=False):
-    if test:
-        fn = lambda x: '_test_' in x
-    else:
-        fn = lambda x: '_test_' not in x
-    result_files = list(filter(fn, os.listdir(path)))
-    all_results_df = pd.Series([], dtype='float32')
-    all_est_names = []
-    for result_file in result_files:
-        result = pkl.load(open(MODEL_COMPARISON_PATH + result_file, 'rb'))
-        all_results_df = pd.concat([all_results_df, result['df']])
-        all_est_names += result['estimators']
-    return {'df': all_results_df, 'estimators': all_est_names}
-
-def viz_comparison(result, dpi=83):
-    '''Plot ROC AUC vs complexity
+def viz_comparison(result: Dict[str, Any]):
+    '''Plot ROC AUC vs complexity, including zoomed-in plot of overlapping region
     '''
-    result_df, result_estimators = result['df'], result['estimators']
-    plt.figure(dpi=dpi)
+    mean_result = result['df']['mean']
+    result_estimators = result['estimators']
+    fig, axes = plt.subplots(nrows=1, ncols=2, figsize=(12, 5))
     for est in np.unique(result_estimators):
         
-        est_result_df = result_df[result_df.index.str.contains(est)]
-        complexities = est_result_df[est_result_df.index.str.contains('Complexity')]
+        est_result_df = mean_result[mean_result.index.str.contains(est)]
+        complexities = est_result_df[est_result_df.index.str.contains('complexity')]
         rocs = est_result_df[est_result_df.index.str.contains('ROC')]
-        complexity_sort_indices = complexities.argsort()
+        complexity_sort_indices = complexities.argsort()    
+        axes[0].plot(complexities[complexity_sort_indices], rocs[complexity_sort_indices], label=est.replace('_', ' '))
+        
+        if est in result['auc_of_auc'].index:
+            area = result['auc_of_auc'][est]
+            label = est.split(' - ')[1] + f' AUC: {area:.3f}'
+            axes[1].plot(complexities[complexity_sort_indices], rocs[complexity_sort_indices], label=label.replace('_', ' '))
 
-        plt.plot(complexities[complexity_sort_indices], rocs[complexity_sort_indices], label=est.replace('_', ' '))
+    for ax in axes:
+        ax.set_xlabel('Complexity score')
+        ax.set_ylabel('Average ROC AUC across comparison datasets')
+        ax.legend(frameon=False, handlelength=1) 
+    axes[1].set_xlim(result['auc_of_auc_lb'], result['auc_of_auc_ub'])
[…]
