+++ tests/notebooks/imodels_comparisons.py
@@ -14,7 +14,8 @@
 # ---
 
 # %%
-import itertools
+import itertools 
+import math
 import os
 import pickle as pkl
 from typing import List, Dict, Any
@@ -22,10 +23,13 @@ from typing import List, Dict, Any
 import numpy as np
 import pandas as pd
 from tqdm import tqdm
-import matplotlib.pyplot as plt
 from sklearn.metrics import accuracy_score, roc_auc_score
 from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
 from sklearn.datasets import fetch_openml
+
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+mpl.rcParams['figure.dpi'] = 250
 
 # %load_ext autoreload
 # %autoreload 2
@@ -47,55 +51,92 @@ datasets = [
         ("vote", 56),
     ]
 
-def get_comparison_result(path, estimator_name, test=False):
+def get_comparison_result(path: str, estimator_name: str, test=False) -> Dict[str, Any]:
     if test:
         result_file = path + 'test/' + f'{estimator_name}_test_comparisons.pkl'
     else:
         result_file = path + 'val/' + f'{estimator_name}_comparisons.pkl'
     return pkl.load(open(result_file, 'rb'))
 
-def viz_comparison(result: Dict[str, Any]):
-    '''Plot ROC AUC vs complexity, including zoomed-in plot of overlapping region
+def get_x_and_y(result_data: pd.Series) -> (pd.Series, pd.Series):
+    complexities = result_data[result_data.index.str.contains('complexity')]
+    rocs = result_data[result_data.index.str.contains('ROC')]
+    complexity_sort_indices = complexities.argsort()    
+    return complexities[complexity_sort_indices], rocs[complexity_sort_indices]
+
+def viz_comparison_val_average(result: Dict[str, Any]) -> None:
+    '''Plot dataset-averaged ROC AUC vs dataset-averaged complexity for different hyperparameter settings
+    of a single model, including zoomed-in plot of overlapping region
     '''
-    mean_result = result['df']['mean']
+    result_data = result['df']['mean']
     result_estimators = result['estimators']
     fig, axes = plt.subplots(nrows=1, ncols=2, figsize=(12, 5))
     for est in np.unique(result_estimators):
         
-        est_result_df = mean_result[mean_result.index.str.contains(est)]
-        complexities = est_result_df[est_result_df.index.str.contains('complexity')]
-        rocs = est_result_df[est_result_df.index.str.contains('ROC')]
-        complexity_sort_indices = complexities.argsort()    
-        axes[0].plot(complexities[complexity_sort_indices], rocs[complexity_sort_indices], label=est.replace('_', ' '))
+        est_result_data = result_data[result_data.index.str.contains(est)]
+        x, y = get_x_and_y(est_result_data)  
+        axes[0].plot(x, y, marker='o', markersize=4, label=est.replace('_', ' '))
         
         if est in result['auc_of_auc'].index:
             area = result['auc_of_auc'][est]
             label = est.split(' - ')[1] + f' AUC: {area:.3f}'
-            axes[1].plot(complexities[complexity_sort_indices], rocs[complexity_sort_indices], label=label.replace('_', ' '))
+            axes[1].plot(x, y, marker='o', markersize=4, label=label.replace('_', ' '))
 
     for ax in axes:
-        ax.set_xlabel('Complexity score')
-        ax.set_ylabel('Average ROC AUC across comparison datasets')
+        ax.set_xlabel('complexity score')
[…]
