### 2021-03-18 5f01346
clean up plotting
