+++ tests/notebooks/imodels_comparisons.py
@@ -6,7 +6,7 @@
 #       extension: .py
 #       format_name: percent
 #       format_version: '1.3'
-#       jupytext_version: 1.10.3
+#       jupytext_version: 1.11.0
 #   kernelspec:
 #     display_name: Python 3
 #     language: python
@@ -14,6 +14,8 @@
 # ---
 
 # %%
+# %load_ext autoreload
+# %autoreload 2
 import itertools 
 import math
 import os
@@ -26,21 +28,24 @@ from tqdm import tqdm
 from sklearn.metrics import accuracy_score, roc_auc_score
 from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
 from sklearn.datasets import fetch_openml
-
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 mpl.rcParams['figure.dpi'] = 250
-
-# %load_ext autoreload
-# %autoreload 2
+import dvu  # for visualization
 
 # change working directory to project root
 if os.getcwd().split('/')[-1] != 'imodels':
     os.chdir('..')
+from notebooks import viz
 
+def get_comparison_result(path: str, estimator_name: str, test=False) -> Dict[str, Any]:
+    if test:
+        result_file = path + 'test/' + f'{estimator_name}_test_comparisons.pkl'
+    else:
+        result_file = path + 'val/' + f'{estimator_name}_comparisons.pkl'
+    return pkl.load(open(result_file, 'rb'))    
+    
 MODEL_COMPARISON_PATH = 'tests/comparison_data/'
-
-# %%
 datasets = [
         ("breast-cancer", 13),
         ("breast-w", 15),
@@ -50,95 +55,6 @@ datasets = [
         ("labor", 4),
         ("vote", 56),
     ]
-
-def get_comparison_result(path: str, estimator_name: str, test=False) -> Dict[str, Any]:
-    if test:
-        result_file = path + 'test/' + f'{estimator_name}_test_comparisons.pkl'
-    else:
-        result_file = path + 'val/' + f'{estimator_name}_comparisons.pkl'
-    return pkl.load(open(result_file, 'rb'))
-
-def get_x_and_y(result_data: pd.Series) -> (pd.Series, pd.Series):
-    complexities = result_data[result_data.index.str.contains('complexity')]
-    rocs = result_data[result_data.index.str.contains('ROC')]
-    complexity_sort_indices = complexities.argsort()    
-    return complexities[complexity_sort_indices], rocs[complexity_sort_indices]
-
-def viz_comparison_val_average(result: Dict[str, Any]) -> None:
-    '''Plot dataset-averaged ROC AUC vs dataset-averaged complexity for different hyperparameter settings
-    of a single model, including zoomed-in plot of overlapping region
-    '''
-    result_data = result['df']['mean']
-    result_estimators = result['estimators']
-    fig, axes = plt.subplots(nrows=1, ncols=2, figsize=(12, 5))
-    for est in np.unique(result_estimators):
-        
-        est_result_data = result_data[result_data.index.str.contains(est)]
-        x, y = get_x_and_y(est_result_data)  
-        axes[0].plot(x, y, marker='o', markersize=4, label=est.replace('_', ' '))
-        
-        if est in result['auc_of_auc'].index:
-            area = result['auc_of_auc'][est]
-            label = est.split(' - ')[1] + f' AUC: {area:.3f}'
-            axes[1].plot(x, y, marker='o', markersize=4, label=label.replace('_', ' '))
-
-    for ax in axes:
-        ax.set_xlabel('complexity score')
-        ax.set_ylabel('ROC AUC')
[…]
