+++ .github/workflows/python-package.yml
@@ -24,6 +24,7 @@ jobs:
         python -m pip install --upgrade pip
         pyversion=$(python -c 'import sys; print(".".join(map(str, sys.version_info[:2])))')
         if [[ $pyversion == '3.6' ]]; then pip install 'numpy<1.20' 'scipy<1.6'; fi
+        pip install git+https://github.com/csinva/data-viz-utils
         pip install . --no-cache-dir
     - name: Test with pytest
       run: |
