+++ .github/workflows/python-package.yml
@@ -22,6 +22,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        pyversion=$(python -c 'import sys; print(".".join(map(str, sys.version_info[:2])))')
+        if [[ $pyversion == '3.6' ]]; then pip install 'numpy<1.20' 'scipy<1.6'; fi
         pip install .
     - name: Test with pytest
       run: |
