### 2021-04-01 12b1889
initial parallel comparison
