### 2021-04-06 9882d25
use rule class in BRS
