+++ experiments/compare_models.py
@@ -9,14 +9,17 @@ import time
 from collections import defaultdict
 from typing import Dict, Any
 
-import imodels
 import numpy as np
 import pandas as pd
+from imodels import (
+    SkopeRulesClassifier as skope, RuleFitClassifier as rfit, FPLassoClassifier as fpl, 
+    FPSkopeClassifier as fps, BayesianRuleListClassifier as brl, GreedyRuleListClassifier as grl,
+    OneRClassifier as oner, BoostedRulesClassifier as brs
+)
 from scipy.interpolate import interp1d
 from scipy.sparse import issparse
 from sklearn.datasets import fetch_openml
-from sklearn.ensemble import GradientBoostingClassifier
-from sklearn.ensemble import RandomForestClassifier
+from sklearn.ensemble import RandomForestClassifier as rf, GradientBoostingClassifier as gb
 from sklearn.metrics import accuracy_score, roc_auc_score
 from sklearn.model_selection import KFold, train_test_split
 from tqdm import tqdm
@@ -45,60 +48,72 @@ LOW_COMPLEXITY_CUTOFF = 30
 # min complexity of curves included in the AUC-of-AUC comparison must be below this value
 MAX_START_COMPLEXITY = 10
 
-BEST_ESTIMATORS = [
-    [('random_forest', RandomForestClassifier(n_estimators=n, max_depth=2)) for n in np.arange(1, 8)],
-    [('gradient_boosting', GradientBoostingClassifier(n_estimators=n, max_depth=1)) for n in np.linspace(1, 20, 10, dtype=int)],
-    [('skope_rules', imodels.SkopeRulesClassifier(n_estimators=n, max_depth=1)) for n in np.linspace(2, 200, 10, dtype=int)],
-    [('rulefit', imodels.RuleFitClassifier(max_rules=n, tree_size=2)) for n in np.linspace(2, 100, 10, dtype=int)],
-    [('fplasso', imodels.FPLassoClassifier(max_rules=n, maxcardinality=1)) for n in np.linspace(2, 100, 10, dtype=int)],
-    [('fpskope', imodels.FPSkopeClassifier(maxcardinality=n, max_depth_duplication=3)) for n in np.arange(1, 5)],
-    [('brl', imodels.BayesianRuleListClassifier(listlengthprior=n, maxcardinality=2)) for n in np.linspace(1, 16, 8)],
-    [('grl', imodels.GreedyRuleListClassifier(max_depth=n)) for n in np.arange(1, 6)],
-    [('oner', imodels.OneRClassifier(max_depth=n)) for n in np.arange(1, 6)],
-    [('brs', imodels.BoostedRulesClassifier(n_estimators=n)) for n in np.linspace(1, 32, 10, dtype=int)]
-]
+
+class Model:
+    def __init__(self, name: str, cls, vary_param: str, vary_param_val: Any, fixed_param: str, fixed_param_val: Any):
+        self.name = name
+        self.cls = cls
+        self.fixed_param = fixed_param
+        self.fixed_param_val = fixed_param_val
+        self.vary_param = vary_param
+        self.vary_param_val = vary_param_val
+        self.kwargs = {self.fixed_param: fixed_param_val, self.vary_param: self.vary_param_val}
+
+
+# BEST_ESTIMATORS = [
+#     [('random_forest', rf(n_estimators=n, max_depth=2)) for n in np.arange(1, 8)],
+#     [('gradient_boosting', gb(n_estimators=n, max_depth=1)) for n in np.linspace(1, 20, 10, dtype=int)],
[…]
