+++ experiments/notebooks/viz.py
@@ -1,40 +1,35 @@
-import itertools 
 import math
-import os
-import pickle as pkl
-from typing import List, Dict, Any
+from typing import List, Dict, Any, Union
 
+import dvu
 import numpy as np
 import pandas as pd
-from tqdm import tqdm
-from sklearn.metrics import accuracy_score, roc_auc_score
-from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
-from sklearn.datasets import fetch_openml
 import matplotlib as mpl
 import matplotlib.pyplot as plt
-mpl.rcParams['figure.dpi'] = 250
-import dvu # for visualization
 dvu.set_style()
+mpl.rcParams['figure.dpi'] = 250
 
-def get_x_and_y(result_data: pd.Series) -> (pd.Series, pd.Series):
-    complexities = result_data[result_data.index.str.contains('complexity')]
-    rocs = result_data[result_data.index.str.contains('ROC')]
-    complexity_sort_indices = complexities.argsort()    
+
+def get_x_and_y(result_data: pd.Series, x_col: str, y_col: str) -> (pd.Series, pd.Series):
+    complexities = result_data[x_col]
+    rocs = result_data[y_col]
+    complexity_sort_indices = complexities.argsort()
     return complexities[complexity_sort_indices], rocs[complexity_sort_indices]
+
 
 def viz_comparison_val_average(result: Dict[str, Any]) -> None:
     '''Plot dataset-averaged ROC AUC vs dataset-averaged complexity for different hyperparameter settings
     of a single model, including zoomed-in plot of overlapping region
     '''
-    result_data = result['df']['mean']
+    result_data = result['df']
     result_estimators = result['estimators']
     fig, axes = plt.subplots(nrows=1, ncols=2, figsize=(12, 5))
     for est in np.unique(result_estimators):
-        
+
         est_result_data = result_data[result_data.index.str.contains(est)]
-        x, y = get_x_and_y(est_result_data)  
+        x, y = get_x_and_y(est_result_data, 'mean_complexity', 'mean_ROCAUC')
         axes[0].plot(x, y, marker='o', markersize=4, label=est.replace('_', ' '))
-        
+
         if est in result['auc_of_auc'].index:
             area = result['auc_of_auc'][est]
             label = est.split(' - ')[1] + f' AUC: {area:.3f}'
@@ -43,7 +38,7 @@ def viz_comparison_val_average(result: Dict[str, Any]) -> None:
     axes[0].set_title('average ROC AUC across all comparison datasets')
     axes[1].set_xlim(result['auc_of_auc_lb'], result['auc_of_auc_ub'])
     axes[1].set_title('Overlapping, low (<30) complexity region only')
-    
+
     for ax in axes:
         ax.set_xlabel('complexity score')
         ax.set_ylabel('ROC AUC')
@@ -51,13 +46,14 @@ def viz_comparison_val_average(result: Dict[str, Any]) -> None:
         dvu.line_legend(fontsize=10, ax=ax)    
     plt.tight_layout()
 
+
 def viz_comparison_test_average(results: List[Dict[str, Any]]) -> None:
     '''Plot dataset-averaged ROC AUC vs dataset-averaged complexity for different models
     '''
     for result in results:
-        mean_result = result['df']['mean']
+        result_data = result['df']
         est = result['estimators'][0]
-        x, y = get_x_and_y(mean_result)  
+        x, y = get_x_and_y(result_data, 'mean_complexity', 'mean_ROCAUC')  
         plt.plot(x, y, marker='o', markersize=2, linewidth=1, label=est.replace('_', ' '))
     plt.xlim(0, 30)
     plt.xlabel('complexity score', size=8)
[…]
