+++ experiments/compare_models.py
@@ -50,28 +50,31 @@ MAX_START_COMPLEXITY = 10
 
 
 class Model:
-    def __init__(self, name: str, cls, vary_param: str, vary_param_val: Any, fixed_param: str, fixed_param_val: Any):
+    def __init__(self, name: str, cls, vary_param: str, vary_param_val: Any, 
+                 fixed_param: str = None, fixed_param_val: Any = None):
         self.name = name
         self.cls = cls
         self.fixed_param = fixed_param
         self.fixed_param_val = fixed_param_val
         self.vary_param = vary_param
         self.vary_param_val = vary_param_val
-        self.kwargs = {self.fixed_param: fixed_param_val, self.vary_param: self.vary_param_val}
+        self.kwargs = {self.vary_param: self.vary_param_val}
+        if self.fixed_param is not None:
+            self.kwargs[self.fixed_param] = self.fixed_param_val
 
 
-# BEST_ESTIMATORS = [
-#     [('random_forest', rf(n_estimators=n, max_depth=2)) for n in np.arange(1, 8)],
-#     [('gradient_boosting', gb(n_estimators=n, max_depth=1)) for n in np.linspace(1, 20, 10, dtype=int)],
-#     [('skope_rules', skope(n_estimators=n, max_depth=1)) for n in np.linspace(2, 200, 10, dtype=int)],
-#     [('rulefit', rfit(max_rules=n, tree_size=2)) for n in np.linspace(2, 100, 10, dtype=int)],
-#     [('fplasso', fpl(max_rules=n, maxcardinality=1)) for n in np.linspace(2, 100, 10, dtype=int)],
-#     [('fpskope', fps(maxcardinality=n, max_depth_duplication=3)) for n in np.arange(1, 5)],
-#     [('brl', brl(listlengthprior=n, maxcardinality=2)) for n in np.linspace(1, 16, 8)],
-#     [('grl', grl(max_depth=n)) for n in np.arange(1, 6)],
-#     [('oner', oner(max_depth=n)) for n in np.arange(1, 6)],
-#     [('brs', brs(n_estimators=n)) for n in np.linspace(1, 32, 10, dtype=int)]
-# ]
+BEST_ESTIMATORS = [
+    [Model('random_forest', rf, 'n_estimators', n, 'max_depth', 1) for n in np.arange(1, 15)],
+    [Model('gradient_boosting', gb, 'n_estimators', n, 'max_depth', 1) for n in np.linspace(1, 20, 10, dtype=int)],
+    [Model('skope_rules', skope, 'n_estimators', n, 'max_depth', 1) for n in np.linspace(2, 200, 10, dtype=int)],
+    [Model('rulefit', rfit, 'max_rules', n, 'tree_size', 2) for n in np.linspace(2, 100, 10, dtype=int)],
+    [Model('fplasso', fpl, 'max_rules', n, 'maxcardinality', 1) for n in np.linspace(2, 100, 10, dtype=int)],
+    [Model('fpskope', fps, 'maxcardinality', n, 'max_depth_duplication', 3) for n in np.arange(1, 5)],
+    [Model('brl', brl, 'listlengthprior', n, 'maxcardinality', 2) for n in np.linspace(1, 16, 8)],
+    [Model('grl', grl, 'max_depth', n) for n in np.arange(1, 6)],
+    [Model('oner', oner, 'max_depth', n) for n in np.arange(1, 6)],
+    [Model('brs', brs, 'n_estimators', n) for n in np.linspace(1, 32, 10, dtype=int)]
+]
 
 ALL_ESTIMATORS = []
 ALL_ESTIMATORS.append(
@@ -106,9 +109,9 @@ ALL_ESTIMATORS.append(
[…]
