+++ imodels/rule_list/greedy_rule_list.py
@@ -4,8 +4,9 @@
 import math
 import numpy as np
 from copy import deepcopy
-from sklearn.base import BaseEstimator, ClassifierMixin, RegressorMixin
-from sklearn.utils.multiclass import check_classification_targets
+
+from sklearn.base import BaseEstimator, ClassifierMixin
+from sklearn.utils.multiclass import check_classification_targets, unique_labels
 from sklearn.utils.validation import check_X_y, check_array, check_is_fitted
 
 from imodels.rule_list.rule_list import RuleList
@@ -110,6 +111,7 @@ class GreedyRuleListClassifier(BaseEstimator, RuleList, ClassifierMixin):
             self.depth += 1  # increase the depth since we call fit once
             self.rules_ = par_node
             self.complexity_ = len(self.rules_)
+            self.classes_ = unique_labels(y)
             return par_node
 
     def predict_proba(self, X):
