### 2021-04-13 0c0ed99
Initial commit for SlipperRule
