+++ imodels/rule_set/slipper.py
@@ -1,0 +1,206 @@
+import pandas as pd
+import numpy as np
+import random
+import string
+
+from copy import deepcopy
+
+from sklearn.base import BaseEstimator, ClassifierMixin
+from sklearn.model_selection import train_test_split
+
+from imodels.util.rule import Rule
+
+
+class SlipperRule(BaseEstimator, ClassifierMixin):
+    """ An estimator that supports building rules as described in
+    A Simple, Fast, and Effective Rule Learner (1999). Intended to be used
+    as part of the BoostedRulesClassifier.
+    """
+
+    def __init__(self, D):
+        self.Z = None
+        self.rule = Rule('')
+        self.D = D
+        self._place_holders = None
+
+    def _make_candidate(self, X, y, curr_rule, feat, A_c):
+        """ Make candidate rules to explore best option
+        greedily
+
+        Parameters
+        ----------
+          X (np.array): Data matrix
+          y (np.array): Response matrix
+          D (np.array): Array of distributions
+        """
+
+        # make candidate rules
+        candidates = []
+        for operator in ['>', '<', '==']:
+            if curr_rule.rule == '':
+                temp_rule = Rule(
+                    str(feat) + ' ' +
+                    operator + ' ' + str(A_c)
+                )
+            else:
+                temp_rule = Rule(
+                    curr_rule.rule + 'and ' + str(feat) +
+                    ' ' + operator + ' ' + str(A_c)
+                )
+            candidates.append(temp_rule)
+
+        # pick best condition
+        Zs = [self._grow_rule_obj(X, y, r) for r in candidates]
+        return candidates[Zs.index(max(Zs))]
+
+    def _rule_predict(self, X, rule): 
+        preds = np.zeros((X.shape[0],))
+        df = pd.DataFrame(X)
+        idx = df.query(rule.rule).index[0]
+        preds[idx] = 1
+        return preds
+
+    def _get_design_matrices(self, X, y, rule):
+        preds = self._rule_predict(X, rule)
+
+        W_plus_idx = np.where((preds == 1) & (y == 1))
+        W_minus_idx = np.where((preds == 1) & (y == 0))
+
+        return np.sum(self.D[W_plus_idx]), np.sum(self.D[W_minus_idx])
+
+    def _grow_rule_obj(self, X, y, rule):
+        W_plus, W_minus = self._get_design_matrices(X, y, rule) 
+        C_R = self.sample_weight(W_plus, W_minus)
+        return np.sqrt(W_plus) - np.sqrt(W_minus)
+
+    def sample_weight(self, plus, minus):
+        """ Calculate learner sample weight
+        in paper this is C_R, which is confidence of learner 
+        """
+        return 0.5 * np.log((plus + (1 / (2 * len(self.D)))) /
+                            (minus + 1 / (2 * len(self.D))))
+
+    def _grow_rule(self, X, y):
+        """ Starts with empty conjunction of conditions and
+        greddily adds rules to mazimize Z_tilda
+
+        Parameters
+        ----------
+        X : array-like, shape (n_samples, n_features)
+            Training vector, where n_samples is the number of samples and
+            n_features is the number of features.
+
+        y : array-like, shape (n_samples,)
+            Target vector relative to X. Has to follow the convention 0 for
+            normal data, 1 for anomalies.
+
+        tol: tolerance for when to end adding conditions to rule
+
+        con_tol: condition tolerance for when to stop
+                 tuning condition for feature
+        """
+
+        stop_condition = False
+        features = list(range(X.shape[1]))
+        curr_rule = Rule('')
+
+        while not stop_condition:
+            candidate_rule = deepcopy(curr_rule)
+            for feat in features:
[…]
