+++ experiments/submit_comparisons.py
@@ -28,7 +28,7 @@ if not args.test_only:
     for i in range(26):
         job_id = s.run(f'python experiments/compare_models.py --model brl --parallel_id {i}')
         brl_job_ids.append(job_id)
-    s.run(f'python experiments/compare_models.py --combine --model brl', depends_on=brl_job_ids)
+    s.run(f'python experiments/combine.py --model brl', depends_on=brl_job_ids)
 
 if not args.val_only:
     models += ['grl', 'oner', 'brs']
@@ -40,4 +40,4 @@ if not args.val_only:
     for i in range(8):
         job_id = s.run(f'python experiments/compare_models.py --test --model brl --parallel_id {i}')
         brl_job_ids.append(job_id)
-    s.run(f'python experiments/compare_models.py --combine --test --model brl', depends_on=brl_job_ids)
+    s.run(f'python experiments/combine.py --test --model brl', depends_on=brl_job_ids)
