+++ experiments/combine.py
@@ -1,14 +1,14 @@
 import argparse
 import glob
 import os
-import pkl
+import pickle as pkl
 
 import pandas as pd
 
 from experiments.compare_models import compute_auc_of_auc, MODEL_COMPARISON_PATH
 
 
-def combine_comparisons(path, model):
+def combine_comparisons(path, model, test):
     all_files = glob.glob(path + '*')
     model_files = list(filter(lambda x: (model in x) and ('comparisons_' in x), all_files))
     model_files_sorted = sorted(model_files, key=lambda x: int(x.split('_')[-1][:-4]))
@@ -23,7 +23,8 @@ def combine_comparisons(path, model):
         'metrics': results_sorted[0]['metrics'],
         'df': df,
     }
-    compute_auc_of_auc(output_dict, column='mean_PRAUC')
+    if not test:
+        compute_auc_of_auc(output_dict, column='mean_PRAUC')
 
     combined_filename = '.'.join(model_files_sorted[0].split('_0.'))
     pkl.dump(output_dict, open(combined_filename, 'wb'))
@@ -35,8 +36,9 @@ def combine_comparisons(path, model):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument('--model', type=str, default=None)
+    parser.add_argument('--test', action='store_true')
     args = parser.parse_args()
 
     path = MODEL_COMPARISON_PATH
     path += 'test/' if args.test else 'val/'
-    combine_comparisons(path, args.model)
+    combine_comparisons(path, args.model, args.test)
