### 2021-04-14 2102b93
remove 3.9-only syntax
