+++ tests/test_notebooks.py
@@ -2,6 +2,7 @@ from unittest.mock import patch, MagicMock
 
 import matplotlib.pyplot
 import dvu
+from sklearn.datasets._openml import OpenMLError
 
 @patch('matplotlib.pyplot')
 @patch('sklearn.tree.plot_tree')
@@ -12,5 +13,8 @@ def test_imodels_demo(mock_pd_style, mock_plot_tree, mock_pyplot):
 @patch('matplotlib.pyplot')
 @patch('dvu.line_legend')
 def test_imodels_comparisons(mock_line_legend, mock_pyplot):
-    mock_pyplot.subplots.return_value = (MagicMock(), MagicMock())
-    from .notebooks import imodels_comparisons
+    try:
+        mock_pyplot.subplots.return_value = (MagicMock(), MagicMock())
+        from .notebooks import imodels_comparisons
+    except OpenMLError:
+        pass
