### 2021-04-19 #3192 @whedon
[REVIEW]: imodels: a python package for fitting interpretable models
**Submitting author:** @csinva (<a href="http://orcid.org/0000-0003-0318-2340">Chandan Singh</a>)
**Repository:** <a href="https://github.com/csinva/imodels" target ="_blank">https://github.com/csinva/imodels</a>
**Version:** 0.2.8
