### 2021-04-19 #3192 @yxoos
Reviewer: @yxoos
Feedback is provided as below:

##	Conflict of interest
- [x] I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
- [x] I confirm that I read and will adhere to the JOSS code of conduct.
- [x]  Repository: Is the source code for this software available at the repository url?
The source code is available on the given URL.
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
Yes, the repository contains an MIT License.
- [x] Contribution and authorship: Has the submitting author (@csinva) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
Yes, the author has made the most contribution to this project.
- [x] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines.
Yes, It does meet the scope.

Installation is seamless on Python 3.6-3.9.
The functional claims are confirmed, and the software works as expected.
There are no performance claims.

I believe if there was more elaboration of the problem being solved, prior to providing an example, a single paragraph of few sentences, it would improve the quality of the software (probably the summary of the submitted manuscript could be used).
I think if the authors started the readme with that then showed the example, as provided, followed by the installation section (highlighted as a section), it would improve the readability of the readme file.
Yes, it is mentioned and handled by installation as well.

Yes, it is provided.
Full documentation is provided for functions.
- [x] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
Yes, there are google colab and Jupyter notebook files included.
Yes, this is clearly stated in the readme file.
- [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
Verified, complete and concise.
- [x] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
Yes, the statement of need is provided and explains the problem that needs to be solved.
- [x] State of the field: Do the authors describe how this software compares to other commonly-used packages?
Yes, there is a short literature review provided.
- [x] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
The quality of writing is coherent and easy to read.
- [x] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
Yes the references are complete and cited properly

Find the docx. document for this review:
[Reviewer_yxoos_feedback.docx](https://github.com/openjournals/joss-reviews/files/6337417/Reviewer_yxoos_feedback.docx)
.
