### 2021-04-23 5bce5b9
clarify installation
