### 2021-04-23 #3192 @csinva
Thanks to both reviewers for their time and thoughtful comments! Glad to hear you feel the package addresses a worthy problem and works smoothly 😊.

Following a suggestion from both reviewers, we made it clearer how to install the project from pip by putting it in its own section [near the top of the readme](https://github.com/csinva/imodels).

Following another suggestion from @yxoos we also added an introductory paragraph in the readme describing a statement of need for the package -- it is now  much clearer.

Thanks also to @vissarion for quick handling of the editing so far.

Best,
Chandan
