### 2021-04-23 #3192 @vissarion
Thanks @csinva for the update. A few steps before recommending acceptance.

@yxoos it seems that your concerns about the statement of need are addressed. If you are OK with this please click the related box in your review.

@jungtaekkim it seems that there is an unclicked box in your review about community guidelines. Does this raise an issue? If not please click it to have a complete review. I personally found [these guidelines](https://github.com/csinva/imodels/blob/master/docs/contributing.md) (referenced from readme) enough.
