### 2021-04-25 67d4cc9
compute meta_auc for more metrics
