+++ experiments/compare_models.py
@@ -58,27 +58,29 @@ def get_dataset(data_id: Union[int, str], onehot_encode_strings: bool = True) ->
     return np.nan_to_num(X.astype('float32')), y
 
 
-def compute_auc_of_auc(result: Dict[str, Any],
-                       low_complexity_cutoff: int = 30,
-                       max_start_complexity: int = 10,
-                       column: str = 'mean_PRAUC') -> None:
+def compute_meta_auc(result_data: pd.DataFrame,
+                     prefix: str,
+                    #  metric_names: list[str] = ['mean_PRAUC', 'mean_ROCAUC', 'mean_accuracy'],
+                     low_complexity_cutoff: int = 30,
+                     max_start_complexity: int = 10) -> Tuple[pd.DataFrame, Tuple[float]]:
 
     # LOW_COMPLEXITY_CUTOFF: complexity score under which a model is considered interpretable
     # MAX_START_COMPLEXITY: min complexity of curves included in the AUC-of-AUC comparison must be below this value
 
-    result_data = result['df']
-    estimators = np.unique(result['estimators'])
+    x_column = f'{prefix}_mean_complexity'
+    compute_columns = result_data.columns[result_data.columns.str.contains('mean')]
+    estimators = np.unique(result_data.index)
     xs = np.empty(len(estimators), dtype=object)
     ys = xs.copy()
 
     for i, est in enumerate(estimators):
 
         est_result_df = result_data[result_data.index.str.contains(est)]
-        complexities_unsorted = est_result_df['mean_complexity']
+        complexities_unsorted = est_result_df[x_column]
         complexity_sort_indices = complexities_unsorted.argsort()
         complexities = complexities_unsorted[complexity_sort_indices]
 
-        roc_aucs = est_result_df[column][complexity_sort_indices]
+        roc_aucs = est_result_df.iloc[complexity_sort_indices][compute_columns]
         xs[i] = complexities.values
         ys[i] = roc_aucs.values
 
@@ -87,29 +89,28 @@ def compute_auc_of_auc(result: Dict[str, Any],
     xs, ys, estimators = xs[mask], ys[mask], estimators[mask]
 
     # find overlapping complexity region for roc-of-roc comparison
-    auc_of_auc_lb = max([x[0] for x in xs])
+    meta_auc_lb = max([x[0] for x in xs])
     endpts = np.array([x[-1] for x in xs])
-    auc_of_auc_ub = min(endpts[endpts > auc_of_auc_lb])
-    auc_of_auc_ub = min(auc_of_auc_ub, low_complexity_cutoff)
+    meta_auc_ub = min(endpts[endpts > meta_auc_lb])
+    meta_auc_ub = min(meta_auc_ub, low_complexity_cutoff)
 
     # handle non-overlapping curves
-    mask = endpts > auc_of_auc_lb
+    mask = endpts > meta_auc_lb
     xs, ys, estimators = xs[mask], ys[mask], estimators[mask]
 
     # compute AUC of interpolated curves in overlap region
-    auc_of_aucs = []
+    meta_aucs = defaultdict(lambda:[])
     for i in range(len(xs)):
+        for c, col in enumerate(compute_columns):
+            f_curve = interp1d(xs[i], ys[i][:, c])
+            x_interp = np.linspace(meta_auc_lb, meta_auc_ub, 100)
+            y_interp = f_curve(x_interp)
+            meta_aucs[col + '_auc'].append(np.trapz(y_interp, x=x_interp))
 
[…]
