### 2021-04-26 640aa20
add simple discretizer
