### 2021-04-26 3a65fd6
accept discretizer kwargs
