+++ imodels/rule_set/fpskope.py
@@ -17,6 +17,7 @@ class FPSkopeClassifier(SkopeRulesClassifier):
                  minsupport=0.1,
                  maxcardinality=2,
                  disc_strategy='mdlp',
+                 disc_kwargs={},
                  verbose=False,
                  precision_min=0.5,
                  recall_min=0.01,
@@ -46,6 +47,7 @@ class FPSkopeClassifier(SkopeRulesClassifier):
                          random_state,
                          verbose)
         self.disc_strategy = disc_strategy
+        self.disc_kwargs = disc_kwargs
         self.minsupport = minsupport
         self.maxcardinality = maxcardinality
         self.verbose = verbose
@@ -62,6 +64,7 @@ class FPSkopeClassifier(SkopeRulesClassifier):
                                     maxcardinality=self.maxcardinality,
                                     undiscretized_features=self.undiscretized_features,
                                     disc_strategy=self.disc_strategy,
+                                    disc_kwargs=self.disc_kwargs,
                                     verbose=self.verbose)[0]
         return [itemsets_to_rules(itemsets)], [np.arange(X.shape[0])], [np.arange(len(self.feature_names))]
 
