### 2021-04-26 #3192 @vissarion
@csinva it seems that we are close to acceptance.

Please make a tagged release and archive, and report the version number and archive DOI here.

Also whedon is complaining about a missing DOI, the proposed DOI is from a different book and I cannot find a DOI for the book you refer to so I think it is OK.

A small comment on the readability of your paper is that you can use the same single sign to indicate equal contribution for both first two authors e.g. mark both with `*`.
