+++ docs/paper/compile_paper.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )" # path to current file
-repodir=$(builtin cd dir; pwd)
+repodir=$(builtin cd dir; cd ..; pwd)
 #repodir=$(builtin cd $docsdir; pwd)
 echo $repodir
 docker run --rm \
