### 2021-04-26 #3192 @csinva
Thanks @vissarion!

Made the [tagged release](https://github.com/csinva/imodels/releases/tag/v0.2.8) and archive on zenodo. The version is 0.2.8 and the doi is 10.5281/zenodo.4026886.

Indeed, I cannot find a doi for that reference either.

I have been unable to get the JOSS template to use the same character for equal contribution - it seems this is a [known issue](https://github.com/openjournals/joss/issues/337).

Please let me know if there's any other way I can help!

Best,

Chandan
