### 2021-04-27 9b3e612
make alpha selection default but optional
