+++ imodels/util/score.py
@@ -6,7 +6,8 @@ import numpy as np
 from sklearn.utils import indices_to_mask
 from sklearn.linear_model import Lasso, LogisticRegression
 from sklearn.linear_model._coordinate_descent import _alpha_grid
-from sklearn.model_selection import KFold
+from sklearn.model_selection import KFold, cross_val_score
+from sklearn.metrics import accuracy_score
 
 from imodels.util.rule import Rule
 
@@ -68,66 +69,80 @@ def _eval_rule_perf(rule: str, X, y) -> Tuple[float, float]:
 
 
 def score_linear(X, y, rules: List[str], 
-                 alphas=None, 
-                 cv=3,
                  penalty='l1',
                  prediction_task='regression',
-                 max_rules=2000, 
+                 max_rules=2000,
+                 alpha=None,
                  random_state=None) -> Tuple[List[Rule], List[float], float]:
-    if alphas is None:
-        if prediction_task == 'regression':
-            alphas = _alpha_grid(X, y)
-        elif prediction_task == 'classification':
-            alphas = [1 / alpha for alpha in np.logspace(-4, 4, num=10, base=10)]
 
-    coef_zero_threshold = 1e-6 / np.mean(np.abs(y))
-    mse_cv_scores = []
-    nonzero_rule_coefs_count = []
-    kf = KFold(cv)
-    
-    # alphas are sorted from most reg. to least reg.
-    for alpha in alphas: 
-        
-        if prediction_task == 'regression':
-            m = Lasso(alpha=alpha, random_state=random_state)
-        else:
-            m = LogisticRegression(penalty=penalty, C=1/alpha, solver='liblinear')
-        mse_cv = 0
-        for train_index, test_index in kf.split(X):
-            X_train, X_test = X[train_index], X[test_index]
-            y_train, y_test = y[train_index], y[test_index]
-            m.fit(X_train, y_train)
-            mse_cv += np.mean((m.predict(X_test) - y_test) ** 2)
-        
-        m.fit(X, y)
-        
-        rule_count = np.sum(np.abs(m.coef_.flatten()) > coef_zero_threshold)
-        if rule_count > max_rules:
-            break
-        nonzero_rule_coefs_count.append(rule_count)
-        mse_cv_scores.append(mse_cv / cv)
-    
-    # rare case in which diff alphas lead to identical scores
-    if np.all(mse_cv_scores == mse_cv_scores[0]):
-        best_alpha = alphas[len(mse_cv_scores) - 1]
+    if alpha is not None:
+        final_alpha = alpha
     else:
-        best_alpha = alphas[np.argmin(mse_cv_scores)]
+        final_alpha = get_best_alpha_under_max_rules(X, y, rules,
+                                                     penalty=penalty,
+                                                     prediction_task=prediction_task,
+                                                     max_rules=max_rules, 
+                                                     random_state=random_state)
 
     if prediction_task == 'regression':
-        lscv = Lasso(alpha=best_alpha, random_state=random_state, max_iter=2000)
+        lscv = Lasso(alpha=final_alpha, random_state=random_state, max_iter=2000)
     else:
-        lscv = LogisticRegression(penalty=penalty, C=1/best_alpha, solver='liblinear',
+        lscv = LogisticRegression(penalty=penalty, C=1/final_alpha, solver='liblinear',
                                   random_state=random_state, max_iter=200)
     lscv.fit(X, y)
 
     coef_ = lscv.coef_.flatten()
-    coefs = list(coef_[:len(coef_)-len(rules)])
+    coefs = list(coef_[:coef_.shape[0]-len(rules)])
[…]
