### 2021-04-27 a4e0821
faster rule querying
