### 2021-04-27 #3192 @vissarion
@csinva thanks for the feedback!

The list of authors in the paper is not the same as the list in 10.5281/zenodo.4026886

Could you please fix this?
