+++ imodels/rule_set/fplasso.py
@@ -53,7 +53,7 @@ class FPLasso(RuleFit):
 
     def _extract_rules(self, X, y) -> List[str]:
         itemsets = extract_fpgrowth(X, y,
-                                    feature_labels=self.feature_placeholders,
+                                    feature_names=self.feature_placeholders,
                                     minsupport=self.minsupport,
                                     maxcardinality=self.maxcardinality,
                                     undiscretized_features=self.undiscretized_features,
