### 2021-04-29 8084cd6
rework pipeline for large datasets
