+++ experiments/compare_models.py
@@ -5,24 +5,26 @@ import argparse
 import os
 import pickle as pkl
 import time
+import warnings
 from collections import defaultdict, OrderedDict
 from typing import Any, Callable, List, Dict, Tuple, Union
+
+warnings.filterwarnings("ignore", message="Bins whose width")
 
 import numpy as np
 import pandas as pd
 from scipy.interpolate import interp1d
-from scipy.sparse import issparse
 from sklearn.base import BaseEstimator
-from sklearn.datasets import fetch_openml
 from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
 from sklearn.metrics import accuracy_score, roc_auc_score, average_precision_score, make_scorer
 from sklearn.model_selection import KFold, train_test_split, cross_validate
 from tqdm import tqdm
 
-from experiments.config import (
-    COMPARISON_DATASETS, EASY_DATASETS, MEDIUM_DATASETS, HARD_DATASETS, BEST_ESTIMATORS, ALL_ESTIMATORS
+from experiments.config.config import (
+    EASY_DATASETS, HARD_DATASETS, BEST_ESTIMATORS, ALL_ESTIMATORS, 
+    BEST_EASY_ESTIMATORS, EASY_ESTIMATORS
 )
-from experiments.util import Model, MODEL_COMPARISON_PATH
+from experiments.util import Model, MODEL_COMPARISON_PATH, get_clean_dataset
 
 
 def get_complexity(estimator: BaseEstimator) -> float:
@@ -41,25 +43,8 @@ def get_complexity(estimator: BaseEstimator) -> float:
         return estimator.complexity_
 
 
-def get_dataset(data_id: Union[int, str], onehot_encode_strings: bool = True) -> Tuple[np.array, np.array]:
-    if type(data_id) is int:
-        # load data from OpenML
-        dataset = fetch_openml(data_id=data_id, as_frame=False)
-        X = dataset.data
-        if issparse(X):
-            X = X.toarray()
-        y = (dataset.target[0] == dataset.target).astype(int)
-    elif type(data_id) is str:
-        # load local datasets
-        df = pd.read_csv(data_id)
-        X, y = df.iloc[:, :-1].values, df.iloc[:, -1].values
-    else:
-        raise NotImplementedError
-    return np.nan_to_num(X.astype('float32')), y
-
-
 def compute_meta_auc(result_data: pd.DataFrame,
-                     prefix: str,
+                     prefix: str = '',
                     #  metric_names: list[str] = ['mean_PRAUC', 'mean_ROCAUC', 'mean_accuracy'],
                      low_complexity_cutoff: int = 30,
                      max_start_complexity: int = 10) -> Tuple[pd.DataFrame, Tuple[float]]:
@@ -67,7 +52,8 @@ def compute_meta_auc(result_data: pd.DataFrame,
     # LOW_COMPLEXITY_CUTOFF: complexity score under which a model is considered interpretable
     # MAX_START_COMPLEXITY: min complexity of curves included in the AUC-of-AUC comparison must be below this value
 
-    x_column = f'{prefix}_mean_complexity'
+    # x_column = f'{prefix}_mean_complexity'
+    x_column = f'mean_complexity'
     compute_columns = result_data.columns[result_data.columns.str.contains('mean')]
     estimators = np.unique(result_data.index)
     xs = np.empty(len(estimators), dtype=object)
@@ -85,8 +71,7 @@ def compute_meta_auc(result_data: pd.DataFrame,
         ys[i] = roc_aucs.values
 
     # filter out curves which start too complex
-    mask = list(map(lambda x: min(x) < max_start_complexity, xs))
[…]
