+++ experiments/config/ensemble_config.py
@@ -1,48 +1,157 @@
 import numpy as np
-from imodels import (
-    SkopeRulesClassifier as skope, RuleFitClassifier as rfit, 
-    BayesianRuleListClassifier as brl, BoostedRulesClassifier as brs
-)
+import imodels
 
-from experiments.util import Model, get_best_models_under_complexity
+from experiments.config.config import RULEFIT_KWARGS, FPL_KWARGS, BRL_KWARGS
+from experiments.models.stablelinear import StableLinearClassifier as stbl
+from experiments.util import Model, get_best_model_under_complexity
 
-weak_learners = [('skope_rules', skope), ('rulefit', rfit), ('brl', brl), ('brs', brs)]
-weak_learners_inst = [get_best_models_under_complexity(c, weak_learners, metric='hard_mean_ROCAUC') for c in np.arange(5, 30, 2)]
-stbl_kw = [{'weak_learners': wl_lst} for wl_lst in weak_learners_inst]
-stbl_cs = lambda: enumerate(np.arange(5, 30, 2))
 
-BEST_ENSEMBLES = []
+def get_weak_learner_inst_list(complexity_limits, easy: bool = False):
+    weak_learners = []
+    for complexity_limit in complexity_limits:
+        weak_learners_c = [
+            get_best_model_under_complexity(c=complexity_limit,
+                                            model_name='skope_rules',
+                                            model_cls=imodels.SkopeRulesClassifier,
+                                            curve_param=1,
+                                            kwargs={'precision_min': 0.3 if not easy else 0.5},
+                                            easy=easy),
+            get_best_model_under_complexity(c=complexity_limit,
+                                            model_name='rulefit',
+                                            model_cls=imodels.RuleFitClassifier,
+                                            curve_param=30 if not easy else 13,
+                                            kwargs=RULEFIT_KWARGS,
+                                            easy=easy),
+            get_best_model_under_complexity(c=complexity_limit,
+                                            model_name='fplasso',
+                                            model_cls=imodels.FPLassoClassifier,
+                                            curve_param=1,
+                                            kwargs=FPL_KWARGS,
+                                            easy=easy),
+            # get_best_model_under_complexity(c=complexity_limit,
+            #                                 model_name='brl',
+            #                                 model_cls=imodels.BayesianRuleListClassifier,
+            #                                 curve_param=3 if not easy else 1,
[…]
