+++ experiments/config/config.py
@@ -25,7 +25,20 @@ HARD_DATASETS = [
     ("juvenile", DATASET_PATH + "ICPSR_03986/DS0001/data_clean.csv")
 ]
 
-BEST_ESTIMATORS = []
+RULEFIT_KWARGS = {'random_state': 0, 'max_rules': None, 'include_linear': False}
+FPL_KWARGS = {'disc_strategy': 'simple', 'max_rules': None}
+BRL_KWARGS = {'disc_strategy': 'simple', 'max_iter': 2000}
+
+BEST_ESTIMATORS = [
+    [Model('random_forest', rf, 'n_estimators', n, 'max_depth', 1) for n in np.arange(1, 26, 3)],
+    [Model('gradient_boosting', gb, 'n_estimators', n, 'max_depth', 1) for n in np.arange(1, 26, 3)],
+    [Model('rulefit', rfit, 'n_estimators', n, 'alpha', 30, RULEFIT_KWARGS) for n in [1, 2, 3] + list(np.arange(7, 50, 6))],
+    [Model('skope_rules', skope, 'n_estimators', n, 'max_depth', 1, {'precision_min': 0.3}) for n in np.arange(1, 137, 15)],
+    [Model('fplasso', fpl, 'alpha', a, 'maxcardinality', 1, FPL_KWARGS) for a in np.logspace(1, 2.8, 10)],
+    [Model('fpskope', fps, 'minsupport', n, 'maxcardinality', 1, {'disc_strategy': 'simple', 'precision_min': 0.4}) for n in np.linspace(0.01, 0.5, 10)],
+    [Model('brl', brl, 'listlengthprior', n, 'listwidthprior', 3, BRL_KWARGS) for n in np.arange(1, 16, 2)],
+    [Model('brs', brs, 'n_estimators', n) for n in np.arange(1, 35, 3)]
+]
 
 
 ALL_ESTIMATORS = []
@@ -50,7 +63,6 @@ ALL_ESTIMATORS.append(
     + [Model('skope_rules - max_depth_3_prec_0.4', skope, 'n_estimators', n, 'max_depth', 3, {'precision_min': 0.4}) for n in np.arange(1, 6)]
     + [Model('skope_rules - max_depth_3_prec_0.5', skope, 'n_estimators', n, 'max_depth', 3, {'precision_min': 0.5}) for n in np.arange(1, 6)]
 )
-RULEFIT_KWARGS = {'random_state': 0, 'max_rules': None, 'include_linear': False}
 ALL_ESTIMATORS.append(
     [Model('rulefit - alpha_30', rfit, 'n_estimators', n, 'alpha', 30, RULEFIT_KWARGS) for n in [1, 2, 3] + list(np.arange(7, 50, 6))]
     + [Model('rulefit - alpha_13', rfit, 'n_estimators', n, 'alpha', 13, RULEFIT_KWARGS) for n in np.arange(1, 24, 2)]
@@ -58,7 +70,6 @@ ALL_ESTIMATORS.append(
     + [Model('rulefit - alpha_2', rfit, 'n_estimators', n, 'alpha', 2, RULEFIT_KWARGS) for n in np.arange(1, 11)]
     + [Model('rulefit - alpha_1', rfit, 'n_estimators', n, 'alpha', 1, RULEFIT_KWARGS) for n in np.arange(1, 11)]
 )
-FPL_KWARGS = {'disc_strategy': 'simple', 'max_rules': None}
 ALL_ESTIMATORS.append(
     [Model('fplasso - max_card_1', fpl, 'alpha', a, 'maxcardinality', 1, FPL_KWARGS) for a in np.logspace(1, 2.8, 10)]
[…]

[…]
