### 2021-05-02 2e48fab
include_linear=False for ensemble
