### 2021-05-03 0d132dc
fix brs complexity underestimate
