### 2021-05-03 #3192 @kyleniemeyer
Hi @yxoos, it looks like you have checked off your reviewer checklist and provided comments - so you are done!

I see in your uploaded document of comments that you did leave a comment on the paper's statement of need, which I am copying here for @csinva and @vissarion's sake:
> I believe if there was more elaboration of the problem being solved, prior to providing an example, a single paragraph of few sentences, it would improve the quality of the software (probably the summary of the submitted manuscript could be used).
>
>I think if the authors started the readme with that then showed the example, as provided, followed by the installation section (highlighted as a section), it would improve the readability of the readme file.
