### 2021-05-03 #3192 @csinva
I believe we did update the statement of need (in [this reply](https://github.com/openjournals/joss-reviews/issues/3192#issuecomment-825362740) and the associated commit) and @yxoos found it satisfactory (based on [this reply](https://github.com/openjournals/joss-reviews/issues/3192#issuecomment-826869370)).
