+++ experiments/config/ensemble_config.py
@@ -29,11 +29,17 @@ def get_weak_learner_inst_list(complexity_limits, easy: bool = False):
                                             kwargs=FPL_KWARGS,
                                             easy=easy),
             # get_best_model_under_complexity(c=complexity_limit,
-            #                                 model_name='brl',
-            #                                 model_cls=imodels.BayesianRuleListClassifier,
-            #                                 curve_param=3 if not easy else 1,
-            #                                 kwargs=BRL_KWARGS,
+            #                                 model_name='fpskope',
+            #                                 model_cls=imodels.FPSkopeClassifier,
+            #                                 curve_param=1,
+            #                                 kwargs={'disc_strategy': 'simple', 'precision_min': 0.3},
             #                                 easy=easy),
+            get_best_model_under_complexity(c=complexity_limit,
+                                            model_name='brl',
+                                            model_cls=imodels.BayesianRuleListClassifier,
+                                            curve_param=1,
+                                            kwargs=BRL_KWARGS,
+                                            easy=easy),
             get_best_model_under_complexity(c=complexity_limit,
                                             model_name='brs',
                                             model_cls=imodels.BoostedRulesClassifier,
@@ -49,7 +55,6 @@ stbl_cs = lambda: enumerate(c_limits)
 def get_ensembles_hard(test=False):
     stbl_kw_hard = [{'weak_learners': wl_lst} for wl_lst in get_weak_learner_inst_list(c_limits)]
 
-    kwargs_1 = [{**kw, 'penalty': 'l2', 'alpha': 1000, 'max_rules': None} for kw in stbl_kw_hard]
     kwargs_2 = [{**kw, 'penalty': 'l2', 'alpha': 100, 'max_rules': None} for kw in stbl_kw_hard]
     kwargs_3 = [{**kw, 'penalty': 'l2', 'alpha': 10, 'max_rules': None} for kw in stbl_kw_hard]
     kwargs_4 = [{**kw, 'penalty': 'l2', 'alpha': 1, 'max_rules': None} for kw in stbl_kw_hard]
@@ -58,54 +63,47 @@ def get_ensembles_hard(test=False):
     BEST_ENSEMBLES = []
     BEST_ENSEMBLES += [
         [Model('stbl_l2_mm0', stbl, 'max_complexity', c, 'min_mult', 0, kwargs_4[i]) for i, c in stbl_cs()],
-        [Model('stbl_l2_mm1', stbl, 'max_complexity', c, 'min_mult', 1, kwargs_5[i]) for i, c in stbl_cs()]
[…]
