### 2021-05-04 c8c3063
brl input validation
